+++ src/forFunctional/index.ts
@@ -1,45 +1,19 @@
 import {
-  GetterDescription,
   GetterFunction,
   Observable,
   ReaxInstance,
   ReaxInstanceForFunctional,
 } from '../types';
-import { useLayoutEffect, useState, useRef, useCallback } from 'react';
+import { useLayoutEffect, useState, useRef } from 'react';
 
-function useReaxGetter(
-  getterDescription: GetterDescription,
-  store: Observable,
-  rawGetters: Record<string, GetterDescription>,
-) {
+function useReaxGetter(store: Observable, getterFunction: () => any) {
   const prevValue = useRef(null);
   const forceRender = useState({})[1];
-
-  const buildLocalGetters = useCallback(
-    (getters: Record<string, GetterDescription>, storeValue: any) => {
-      const result: Record<string, any> = {};
-      for (const getterKey in getters) {
-        if (!Object.prototype.hasOwnProperty.call(getters, getterKey)) continue;
-        const getterDesc = getters[getterKey];
-        result[getterKey] = () => getterDesc.func(storeValue, result);
-      }
-      return result;
-    },
-    [],
-  );
-
-  let currentValue = getterDescription.func(
-    store.value[getterDescription.module],
-    buildLocalGetters(rawGetters, store.value[getterDescription.module]),
-  );
+  let currentValue = getterFunction();
 
   useLayoutEffect(() => {
-    const unsubscribe = store.subscribe((v: any) => {
-      const storeValue = v[getterDescription.module];
-      currentValue = getterDescription.func(
-        storeValue,
-        buildLocalGetters(rawGetters, storeValue),
-      );
+    const unsubscribe = store.subscribe(() => {
+      currentValue = getterFunction();
       if (prevValue.current !== currentValue) {
         prevValue.current = currentValue;
         forceRender({});
@@ -51,15 +25,13 @@ function useReaxGetter(
   return currentValue;
 }
 
-const buildGetters = (store: ReaxInstance) =>
-  Object.entries(store.rawGetters).reduce(
+const buildGetters = (
+  store: Observable,
+  getterFunctions: Record<string, () => any>,
+) =>
+  Object.entries(getterFunctions).reduce(
     (total: Record<string, GetterFunction>, [key, func]) => {
-      total[key] = () =>
-        useReaxGetter(
-          func as GetterDescription,
-          store.$$instance,
-          store.rawGetters,
-        );
+      total[key] = () => useReaxGetter(store, func);
       return total;
     },
     {},
@@ -70,13 +42,14 @@ export default function forFunctional(
 ): ReaxInstanceForFunctional {
   Object.defineProperties(store, {
     getters: {
-      value: buildGetters(store),
+      value: buildGetters(store.$$instance, store.$$getterFunctions),
       enumerable: true,
       writable: true,
     },
     $$rebuildGetters: {
       value: function () {
-        this.getters = buildGetters(store);
+        this.$$selfUpdate();
+        this.getters = buildGetters(this.$$instance, this.$$getterFunctions);
       },
     },
   });