+++ src/index.ts
@@ -1,4 +1,5 @@
 import {
+  ActionFunction,
   GetterDescription,
   MutationDescription,
   Observable,
@@ -47,6 +48,13 @@ const createModule = (name: string, descriptor: StoreDescriptor) => {
       },
       {},
     ),
+    moduleActions: Object.entries(descriptor.actions || {}).reduce(
+      (total: Record<string, any>, [key, func]) => {
+        total[prefix + key] = func;
+        return total;
+      },
+      {},
+    ),
   };
 };
 
@@ -60,23 +68,36 @@ const deleteByKeyPart = (startsWith: string, object: any) => {
   }
 };
 
+const createGetters = (
+  getters: Record<string, GetterDescription>,
+  store: Observable,
+) =>
+  Object.entries(getters).reduce(
+    (allGetters: Record<string, () => any>, [getterKey, getterDesc]) => {
+      allGetters[getterKey] = () =>
+        getterDesc.func(store.value[getterDesc.module], allGetters);
+      return allGetters;
+    },
+    {},
+  );
+
 export function createStore(descriptor: StoreDescriptor): ReaxInstance {
   if (!Object.prototype.hasOwnProperty.call(descriptor, 'state'))
     throw new ReferenceError('[reax] state must be defined');
   let state: any = {};
   let mutations: Record<string, MutationDescription> = {};
   let rawGetters: Record<string, GetterDescription> = {};
+  let allActions: Record<string, ActionFunction> = {};
   const { modules = {}, ...root } = descriptor;
   const allModules = { root, ...modules };
 
   const attachModule = (name: string, descriptor: StoreDescriptor) => {
-    const { moduleState, moduleGetters, moduleMutations } = createModule(
-      name,
-      descriptor,
-    );
-    state = { ...state, ...moduleState };
-    mutations = { ...mutations, ...moduleMutations };
-    rawGetters = { ...rawGetters, ...moduleGetters };
+    const { moduleState, moduleGetters, moduleMutations, moduleActions } =
+      createModule(name, descriptor);
+    state = Object.assign(state, moduleState);
+    mutations = Object.assign(mutations, moduleMutations);
+    rawGetters = Object.assign(rawGetters, moduleGetters);
+    allActions = Object.assign(allActions, moduleActions);
     return moduleState;
   };
 
@@ -86,24 +107,22 @@ export function createStore(descriptor: StoreDescriptor): ReaxInstance {
 
   const storeState = createObservable(state);
 
-  return {
-    $$instance: storeState,
+  const reaxStore = {
     get state() {
       const { root, ...rest } = storeState.value;
-      return { ...root, ...rest };
+      return Object.assign(root, rest);
     },
     commit: (type: string, payload: any) => {
       if (!mutations[type])
         return console.error('[reax] unknown mutation type:', type);
       const { func, module } = mutations[type];
-      const _st = { ...storeState.value[module] };
+      const _st = JSON.parse(JSON.stringify(storeState.value[module]));
       func(_st, payload);
       storeState.value = { ...storeState.value, [module]: { ..._st } };
     },
-    subscribe: storeState.subscribe.bind(storeState),
     registerModule: function (name: string, descriptor: StoreDescriptor) {
       const moduleState = attachModule(name, descriptor);
-      this.$$rebuildGetters && this.$$rebuildGetters();
+      this.$$rebuildGetters ? this.$$rebuildGetters() : this.$$selfUpdate();
       storeState.value = { ...storeState.value, ...moduleState };
     },
     unregisterModule: function (name: string) {
@@ -111,9 +130,31 @@ export function createStore(descriptor: StoreDescriptor): ReaxInstance {
       deleteByKeyPart(name, mutations);
       deleteByKeyPart(name, this.getters);
       storeState.value = newState;
+      this.$$rebuildGetters ? this.$$rebuildGetters() : this.$$selfUpdate();
     },
-    get rawGetters() {
-      return rawGetters;
+    dispatch: async (type: string, payload?: any) => {
+      return allActions[type](
+        {
+          commit: reaxStore.commit,
+          dispatch: reaxStore.dispatch,
+          getters: reaxStore.$$getterFunctions,
+        },
+        payload,
+      );
+    },
+    $$getterFunctions: createGetters(rawGetters, storeState),
+    $$subscribe: storeState.subscribe.bind(storeState),
+    $$instance: storeState,
+    $$selfUpdate: function () {
+      this.$$getterFunctions = createGetters(rawGetters, storeState);
     },
   };
+
+  Object.keys(reaxStore).forEach(
+    (key: string) =>
+      key.startsWith('$$') &&
+      Object.defineProperty(reaxStore, key, { enumerable: false }),
+  );
+
+  return reaxStore;
 }