+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 1.0.0
+* BREAKING: `subscribe` => `$$subscribe`;
+* BREAKING: `rawGetters` deleted (replaced with `$$getterFunctions`);
+* actions implemented;
+* getter system simplified.
+
 ### 0.1.0
 * Add optional `getters` argument to all getters functions
   (for accessing getters inside getter);